mediaapi/fileutils: Move fileutils to its own package
